pyproject.toml
[project]
name = "kdo-video"
version = "1.0.0"
description = "Video metadata tagger with web UI for NAS"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "ffmpeg-python>=0.2.0",
    "ultralytics>=8.1.0",
    "opencv-python>=4.9.0",
    "supervision>=0.21.0",
    "python-multipart>=0.0.6",
    "openpyxl>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true

[tool.fastapi]
reload = true
